+++ src/plugins/api.ts
@@ -75,6 +75,7 @@ class ApiPlugin extends deploy_objects.DeployPluginBase {
 
             if (showPasswordPrompt) {
                 vscode.window.showInputBox({
+                    ignoreFocusOut: true,
                     placeHolder: i18.t('prompts.inputPassword'),
                     password: true,
                 }).then((passwordFromUser) => {